+++ contracts/src/access/control.rs
@@ -50,6 +50,7 @@ use stylus_sdk::{
     prelude::*,
     storage::{StorageBool, StorageFixedBytes, StorageMap},
 };
+use crate::utils::introspection::erc165::{Erc165, IErc165};
 
 #[cfg_attr(coverage_nightly, coverage(off))]
 mod sol {
@@ -400,11 +401,19 @@ impl AccessControl {
     }
 }
 
+impl IErc165 for AccessControl {
+    fn supports_interface(interface_id: FixedBytes<4>) -> bool {
+        <Self as IAccessControl>::INTERFACE_ID == u32::from_be_bytes(*interface_id) || 
+        Erc165::supports_interface(interface_id)
+    }
+}
+
 #[cfg(all(test, feature = "std"))]
 mod tests {
-    use alloy_primitives::Address;
+    use alloy_primitives::{Address, FixedBytes};
     use motsu::prelude::Contract;
     use stylus_sdk::prelude::TopLevelStorage;
+    use crate::utils::introspection::erc165::IErc165;
 
     use super::{AccessControl, Error, IAccessControl};
 
@@ -709,5 +718,24 @@ mod tests {
         let role_revoked =
             contract.sender(alice)._revoke_role(ROLE.into(), bob);
         assert!(!role_revoked);
+    }
+
+    #[motsu::test]
+    fn interface_id() {
+        let actual = <AccessControl as IAccessControl>::INTERFACE_ID;
+        let expected = 0x7965db0b; 
+        assert_eq!(actual, expected);
+    }
+
+    #[motsu::test]
+    fn supports_interface() {
+        assert!(AccessControl::supports_interface(
+            <AccessControl as IAccessControl>::INTERFACE_ID.into()
+    ));
+        assert!(AccessControl::supports_interface(
+            <AccessControl as IErc165>::INTERFACE_ID.into()
+    ));
+        let fake_interface_id = 0x12345678u32;
+        assert!(!AccessControl::supports_interface(fake_interface_id.into()));
     }
 }