+++ contracts/src/token/erc20/extensions/erc4626.rs
@@ -9,7 +9,9 @@
 
 use alloc::{vec, vec::Vec};
 
-use alloy_primitives::{uint, Address, U256, U8};
+use alloy_primitives::{uint, Address, U256, U8, FixedBytes};
+use crate::utils::introspection::erc165::{Erc165, IErc165};
+
 pub use sol::*;
 use stylus_sdk::{
     call::{Call, MethodError},
@@ -1150,15 +1152,23 @@ impl Erc4626 {
     }
 }
 
+impl IErc165 for Erc4626 {
+    fn supports_interface(interface_id: FixedBytes<4>) -> bool {
+        <Self as IErc4626>::INTERFACE_ID == u32::from_be_bytes(*interface_id) ||
+        Erc165::supports_interface(interface_id)
+    }
+}
+
 // TODO: Add missing tests once `motsu` supports calling external contracts.
 #[cfg(all(test, feature = "std"))]
 mod tests {
-    use alloy_primitives::{address, Address, U256, U8};
+    use alloy_primitives::{address, Address, U256, U8, FixedBytes};
     use motsu::prelude::Contract;
     use stylus_sdk::prelude::*;
 
     use super::{Erc4626, IErc4626};
     use crate::token::erc20::Erc20;
+    use crate::utils::introspection::erc165::IErc165;
 
     #[storage]
     struct Erc4626TestExample {
@@ -1243,5 +1253,25 @@ mod tests {
 
         let decimals = contract.sender(alice).erc4626.decimals();
         assert_eq!(decimals, underlying_decimals + new_decimal_offset);
+    }
+
+
+    #[motsu::test]
+    fn erc4626_interface_id() {
+        let actual = <Erc4626 as IErc4626>::INTERFACE_ID;
+        let expected = 0x4d49e87d; 
+        assert_eq!(actual, expected);
+    }
+
+    #[motsu::test]
+    fn erc4626_supports_interface() {
+        assert!(Erc4626::supports_interface(
+            <Erc4626 as IErc4626>::INTERFACE_ID.into()
+    ));
+        assert!(Erc4626::supports_interface(
+            <Erc4626 as IErc165>::INTERFACE_ID.into()
+    ));
+        let fake_interface_id = 0x12345678u32;
+        assert!(!Erc4626::supports_interface(fake_interface_id.into()));
     }
 }