+++ contracts/src/token/erc20/utils/safe_erc20.rs
@@ -11,7 +11,9 @@
 
 use alloc::{vec, vec::Vec};
 
-use alloy_primitives::{Address, U256};
+use alloy_primitives::{Address, U256, FixedBytes};
+use crate::utils::introspection::erc165::{Erc165, IErc165};
+
 use alloy_sol_types::SolCall;
 pub use sol::*;
 use stylus_sdk::{
@@ -384,9 +386,17 @@ impl SafeErc20 {
     }
 }
 
+impl IErc165 for SafeErc20 {
+    fn supports_interface(interface_id: FixedBytes<4>) -> bool {
+        Erc165::supports_interface(interface_id)
+    }
+}
+
 #[cfg(all(test, feature = "std"))]
 mod tests {
     use super::SafeErc20;
+    use crate::utils::introspection::erc165::IErc165;
+    
     #[test]
     fn encodes_true_empty_slice() {
         assert!(!SafeErc20::encodes_true(&[]));
@@ -415,5 +425,17 @@ mod tests {
     #[test]
     fn encodes_true_wrong_bytes() {
         assert!(!SafeErc20::encodes_true(&[0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 1]));
+    }
+
+    #[motsu::test]
+    fn safe_erc20_supports_interface() {
+        assert!(SafeErc20::supports_interface(
+            <SafeErc20 as IErc165>::INTERFACE_ID.into()
+    ));
+            assert!(SafeErc20::supports_interface(
+                <SafeErc20 as ISafeErc20>::INTERFACE_ID.into()
+            ));
+        let fake_interface_id = 0x12345678u32;
+        assert!(!SafeErc20::supports_interface(fake_interface_id.into()));
     }
 }